Result targetScore Double로 변경,ScoreType 에 따라 다른 SQL 로직 구현

+++ src/main/java/com/example/jsgamesbackendmain/Model/DAO/GameDAO.java
@@ -18,7 +18,7 @@ public class GameDAO {
     private Long userId;
     private String imageUrl;
     private String gameUrl;
-    private Long targetScore;
+    private Double targetScore;
     private ScoreType scoreType;
     private String description;
 }
