+++ src/main/java/com/example/jsgamesbackendmain/Model/DTO/Game/GameDTO.java
@@ -13,7 +13,7 @@ public class GameDTO {
     private Long userId;
     private String imageUrl;
     private String gameUrl;
-    private Long targetScore;
+    private Double targetScore;
     private ScoreType scoreType;
     private String description;
